I made the conversions from chess to matrix and from matrix to chess inlined + makefile organization
I made the funciotions to convert from the matrix format to the chess format and from the chess format to the matrix format (suitable to the playing collor).
I've moved the implementation from the matrix_transformation.cpp file to the matrix_transformation.h file to use the always inline atribute.
I organized a litle the makefile.