A lot has changed
The variable name color changed from collor, again
Finished the init part in communications.cpp. TODO make a string with all the features we want to enable and give them to the engine and listen for the reply from the engine and sincronise the number of couts.
Pieces.cpp -> made the move part. needs to be done more gently and cleaner, get rid of the search and look nicer. For now I think it is enough. TODO the check_validity function needs to be done and the get_piece_directions get finished. Also I changed some variables to char type for memory savings.
main.cpp -> made a test with hardcoded premoves and when remains out of moves, it resigns. I need to check if a move is aviable to not make illegal moves with the hardcoded premoves.