Made a game class and made the search for mooves for rook and bishop external
I made a game object that holds the info for the player and for the enemy. This will come in handy at the minimax part.
I added general move directions in pieces.h that will be used to find the possible moves for the rook, bishop.
Also made functions that put the mooves for the rook and the bishop to be more modular and organized. I didn't change too much stuff as we have to meet and resolve the merge conflicts beforehand. (we need to include the pieces.cpp from the folder to pieces.cpp that is in the root of the project.

Co-Authored-By: Oprea Olivia <oliviawoow@users.noreply.github.com>
Co-Authored-By: jolie369 <jolie369@users.noreply.github.com>