Finished the "get_moves_x" functions
We decided to get rid of the duplicate code and decided to call the same function with different parameters. So all the logic is in the "get_moves" function now.
"get_moves_x" is just calling the funciton "get_moves" with the right parameterf, if we want our moves or enemy's moves.
We also decided the structure of the vector returned by the just mentioned functions, more preciselly a vector of vectors with the fallowing information: {src_x, src_y, dst_x, dst_y, piece_type, priority_code}

Co-Authored-By: Oprea Olivia <oliviawoow@users.noreply.github.com>