An improvement to the modularity of the code and it's extensions
Now we call the same funciton and if the need arises to change something in it, now we have to change only in one place.
Also I added TODOs to start an heuristic, in short terms we will decide based of hte return value of the "check_validity" function with which moves to start in order to maximize the alpha beta potential to cut on useless subtrees before computing them.
Not all funitons that call "check_validity" are ready to take a return value higher than 2, that has to be changed.

P.S. I added an issue because some functions do not save the returned value of "check_validity"